move over to V2 with errors
