+++ catalogue/webapp/pages/works.js
@@ -1,154 +1,193 @@
 // @flow
-import {Fragment, Component} from 'react';
+import type {Context} from 'next';
+import type {CatalogueApiError, CatalogueResultsList} from '../services/catalogue/works';
+// $FlowFixMe: using react aloha for hooks, which isn't in the typedefs
+import {Fragment, useState, useEffect, useRef} from 'react';
 import Router from 'next/router';
-import NextLink from 'next/link';
 import {font, grid, spacing, classNames} from '@weco/common/utils/classnames';
-import {pageStore} from '@weco/common/views/components/PageWrapper/PageWrapper';
 import PageLayout from '@weco/common/views/components/PageLayout/PageLayout';
 import InfoBanner from '@weco/common/views/components/InfoBanner/InfoBanner';
 import Icon from '@weco/common/views/components/Icon/Icon';
 import SearchBox from '@weco/common/views/components/SearchBox/SearchBox';
 import StaticWorksContent from '@weco/common/views/components/StaticWorksContent/StaticWorksContent';
 import WorkPromo from '@weco/common/views/components/WorkPromo/WorkPromo';
-import Pagination, {PaginationFactory} from '@weco/common/views/components/Pagination/Pagination';
-import type {Props as PaginationProps} from '@weco/common/views/components/Pagination/Pagination';
-import type {GetInitialPropsProps} from '@weco/common/views/components/PageWrapper/PageWrapper';
+import Paginator from '@weco/common/views/components/Paginator/Paginator';
+import ErrorPage from '@weco/common/views/components/ErrorPage/ErrorPage';
 import {getWorks} from '../services/catalogue/works';
-import {getWorks as getWorksV2} from '../services/catalogue/worksv2';
 import {workLink, worksLink} from '../services/catalogue/links';
 
-type PageProps = {|
-  query: ?string,
-  page: ?number,
-  works: {| results: [], totalResults: number |},
-  pagination: ?PaginationProps,
-  version: ?number
-|}
-
-type ComponentProps = {|
-  ...PageProps,
-  handleSubmit: (SyntheticEvent<HTMLFormElement>) => void
+type Props = {|
+  initialQuery: ?string,
+  initialWorks: ?CatalogueResultsList | CatalogueApiError,
+  initialPage: ?number,
+  filters: Object
 |}
 
 export const Works = ({
-  query,
-  page,
-  works,
-  pagination,
-  handleSubmit,
-  version
-}: ComponentProps) => (
-  <PageLayout
-    title={`${query ? `${query} | ` : ''}Catalogue search`}
-    description='Search through the Wellcome Collection image catalogue'
-    url={{pathname: '/works', query: worksLink({query, page}).href}}
-    openGraphType={'website'}
-    jsonLd={{ '@type': 'WebPage' }}
-    imageUrl={null}
-    imageAltText={null}
-  >
-    <InfoBanner text={`Coming from Wellcome Images? All freely available images have now been moved to the Wellcome Collection website. Here we're working to improve data quality, search relevance and tools to help you use these images more easily`} cookieName='WC_wellcomeImagesRedirect' />
+  initialQuery,
+  initialWorks,
+  filters,
+  initialPage
+}: Props) => {
+  if (initialWorks && initialWorks.type === 'Error') {
+    return (
+      <PageLayout
+        title={initialWorks.httpStatus.toString()}
+        description={''}
+        url={{pathname: `/works`}}
+        openGraphType={'website'}
+        jsonLd={{ '@type': 'WebPage' }}
+        oEmbedUrl={`https://wellcomecollection.org/works`}
+        imageUrl={null}
+        imageAltText={null}>
+        <ErrorPage
+          errorStatus={initialWorks.httpStatus}
+        />
+      </PageLayout>
+    );
+  }
 
-    <div className={classNames([
-      'row bg-cream',
-      spacing({s: 3, m: 5}, {padding: ['top']}),
-      spacing({s: 3, m: 4, l: 6}, {padding: ['bottom']})
-    ])}>
-      <div className='container'>
-        <div className='grid'>
-          <div className={grid({s: 12, m: 12, l: 12, xl: 12})}>
-            <div className={classNames([
-              'flex flex--h-space-between flex--v-center flex--wrap',
-              spacing({s: 2}, {margin: ['bottom']})
-            ])}>
-              <h1 className={classNames([
-                font({s: 'WB6', m: 'WB4'}),
-                spacing({s: 2}, {margin: ['bottom']}),
-                spacing({s: 4}, {margin: ['right']}),
-                spacing({s: 0}, {margin: ['top']})
-              ])}>Search our images</h1>
-              <div className='plain-text flex flex--v-center'>
-                <Icon name='underConstruction' extraClasses='margin-right-s2' />
-                <p className='no-margin'>We’re improving how search works. <a href='/works/progress'>Find out more</a>.</p>
+  const [query, setQuery] = useState(initialQuery);
+  const [works, setWorks] = useState(initialWorks);
+  const [page, setPage] = useState(initialPage);
+  const [loading, setLoading] = useState(false);
+  useEffect(() => {
+    document.title = `${query} | Catalogue search | Wellcome Collection`;
+  }, [query]);
+
+  // On the initial render from next, we dont want to run the router, nor update
+  // the works so we just skip this effect for now.
+  // See: https://reactjs.org/docs/hooks-faq.html#is-there-something-like-instance-variables
+  const initialRender = useRef(true);
+  useEffect(() => {
+    if (initialRender.current) {
+      initialRender.current = false;
+      return;
+    }
+
+    // TODO: (flowtype) next's typing says that these need to be string, this isn't true,
+    // you can use URL like objects too
+    Router.push(
+      // $FlowFixMe
+      worksLink({query, page}).href,
+      // $FlowFixMe
+      worksLink({query, page}).as,
+      { shallow: true }
+    );
+
+    if (query && query !== '') {
+      setLoading(true);
+      // TODO: Look into memoiszing results so we don't hit the API again
+      //       See: https://reactjs.org/docs/hooks-reference.html#usememo
+
+      // TODO: Return a cleanup function here to stop the network request.
+      getWorks({query, page, filters}).then(setWorks).then(() => setLoading(false));
+    } else {
+      setWorks(null);
+    }
+  }, [page, query]);
+
+  return (
+    <PageLayout
+      title={`${query ? `${query} | ` : ''}Catalogue search`}
+      description='Search through the Wellcome Collection image catalogue'
+      url={{pathname: '/works', query: worksLink({query, page}).href}}
+      openGraphType={'website'}
+      jsonLd={{ '@type': 'WebPage' }}
+      imageUrl={null}
+      imageAltText={null}
+    >
+      <InfoBanner text={`Coming from Wellcome Images? All freely available images have now been moved to the Wellcome Collection website. Here we're working to improve data quality, search relevance and tools to help you use these images more easily`} cookieName='WC_wellcomeImagesRedirect' />
+
+      <div className={classNames([
+        'row bg-cream',
+        spacing({s: 3, m: 5}, {padding: ['top']}),
+        spacing({s: 3, m: 4, l: 6}, {padding: ['bottom']})
+      ])}>
+        <div className='container'>
+          <div className='grid'>
+            <div className={grid({s: 12, m: 12, l: 12, xl: 12})}>
+              <div className={classNames([
+                'flex flex--h-space-between flex--v-center flex--wrap',
+                spacing({s: 2}, {margin: ['bottom']})
+              ])}>
+                <h1 className={classNames([
+                  font({s: 'WB6', m: 'WB4'}),
+                  spacing({s: 2}, {margin: ['bottom']}),
+                  spacing({s: 4}, {margin: ['right']}),
+                  spacing({s: 0}, {margin: ['top']})
+                ])}>Search our images</h1>
+                <div className='plain-text flex flex--v-center'>
+                  <Icon name='underConstruction' extraClasses='margin-right-s2' />
+                  <p className='no-margin'>We’re improving how search works. <a href='/works/progress'>Find out more</a>.</p>
+                </div>
               </div>
             </div>
           </div>
-        </div>
-        <div className='grid'>
-          <div className={grid({s: 12, m: 10, l: 8, xl: 8})}>
-            <SearchBox
-              action=''
-              id='search-works'
-              name='query'
-              query={query || ''}
-              autofocus={true}
-              onSubmit={handleSubmit} />
-
-            {!query
-              ? <p className={classNames([
-                spacing({s: 4}, {margin: ['top']}),
-                font({s: 'HNL4', m: 'HNL3'})
-              ])}>Find thousands of Creative Commons licensed images from historical library materials and museum objects to contemporary digital photographs.</p>
-              : <p className={classNames([
-                spacing({s: 2}, {margin: ['top', 'bottom']}),
-                font({s: 'LR3', m: 'LR2'})
-              ])}>{works.totalResults !== 0 ? works.totalResults : 'No'} results for &apos;{query}&apos;
-              </p>
-            }
+          <div className='grid'>
+            <div className={grid({s: 12, m: 10, l: 8, xl: 8})}>
+              <SearchBox
+                action=''
+                id='search-works'
+                name='query'
+                query={query || ''}
+                autofocus={true}
+                onSubmit={async (event) => {
+                  event.preventDefault();
+                  const form = event.currentTarget;
+                  // $FlowFixMe
+                  const newQuery = form.elements.query.value;
+                  setQuery(newQuery);
+                  setPage(1);
+                }} />
+              {!works
+                ? <p className={classNames([
+                  spacing({s: 4}, {margin: ['top']}),
+                  font({s: 'HNL4', m: 'HNL3'})
+                ])}>Find thousands of Creative Commons licensed images from historical library materials and museum objects to contemporary digital photographs.</p>
+                : <p className={classNames([
+                  spacing({s: 2}, {margin: ['top', 'bottom']}),
+                  font({s: 'LR3', m: 'LR2'})
+                ])}>{works.totalResults !== 0 ? works.totalResults : 'No'} results for &apos;{query}&apos;
+                </p>
+              }
+            </div>
           </div>
         </div>
       </div>
-    </div>
 
-    {!query &&
-      <StaticWorksContent />
-    }
+      {!query &&
+        <StaticWorksContent />
+      }
 
-    {query &&
-      <Fragment>
-        {pagination && pagination.range &&
+      {works && works.results.length > 0 &&
+        <Fragment>
           <div className={`row ${spacing({s: 3, m: 5}, {padding: ['top']})}`}>
             <div className='container'>
               <div className='grid'>
                 <div className='grid__cell'>
                   <div className='flex flex--h-space-between flex--v-center'>
-
                     <Fragment>
-                      <div className={`flex flex--v-center font-pewter ${font({s: 'LR3', m: 'LR2'})}`}>
-                          Showing {pagination.range.beginning} - {pagination.range.end}
-                      </div>
-                      <Pagination
-                        total={pagination.total}
-                        prevPage={pagination.prevPage}
-                        currentPage={pagination.currentPage}
-                        pageCount={pagination.pageCount}
-                        nextPage={pagination.nextPage}
-                        nextQueryString={pagination.nextQueryString}
-                        prevQueryString={pagination.prevQueryString} />
+                      <Paginator
+                        currentPage={page}
+                        pageSize={works.pageSize}
+                        totalResults={works.totalResults}
+                        link={worksLink({query, page})}
+                        onPageChange={async (event, newPage) => {
+                          event.preventDefault();
+                          setPage(newPage);
+                        }}
+                      />
                     </Fragment>
                   </div>
                 </div>
               </div>
             </div>
           </div>
-        }
-        {version === 2 &&
-          <div className='container'>
-            <div className='grid'>
-              <div className={grid({s: 12})}>
-                {works.results.map(result => (
-                  <NextLink href={`/work?id=${result.id}`} as={`/works/${result.id}`} key={result.id}>
-                    <a>
-                      <h2>{result.title}</h2>
-                    </a>
-                  </NextLink>
-                ))}
-              </div>
-            </div>
-          </div>
-        }
-        {version !== 2 &&
-          <div className={`row ${spacing({s: 4}, {padding: ['top']})}`}>
+
+          <div
+            className={`row ${spacing({s: 4}, {padding: ['top']})}`}
+            style={{ opacity: loading ? 0 : 1 }}>
             <div className='container'>
               <div className='grid'>
                 {works.results.map(result => (
@@ -169,110 +208,76 @@ export const Works = ({
               </div>
             </div>
           </div>
-        }
-        {pagination && pagination.range &&
+
           <div className={`row ${spacing({s: 10}, {padding: ['top', 'bottom']})}`}>
             <div className='container'>
               <div className='grid'>
                 <div className='grid__cell'>
                   <div className='flex flex--h-space-between flex--v-center'>
-
                     <Fragment>
-                      <div className={`flex flex--v-center font-pewter ${font({s: 'LR3', m: 'LR2'})}`}>
-                        Showing {pagination.range.beginning} - {pagination.range.end}
-                      </div>
-                      <Pagination
-                        total={pagination.total}
-                        prevPage={pagination.prevPage}
-                        currentPage={pagination.currentPage}
-                        pageCount={pagination.pageCount}
-                        nextPage={pagination.nextPage}
-                        nextQueryString={pagination.nextQueryString}
-                        prevQueryString={pagination.prevQueryString} />
+                      <Paginator
+                        currentPage={page}
+                        pageSize={works.pageSize}
+                        totalResults={works.totalResults}
+                        link={worksLink({query, page})}
+                        onPageChange={async (event, newPage) => {
+                          event.preventDefault();
+                          setPage(newPage);
+                        }}
+                      />
                     </Fragment>
                   </div>
                 </div>
               </div>
             </div>
           </div>
-        }
-      </Fragment>
-    }
-  </PageLayout>
-);
+        </Fragment>
+      }
 
-export class WorksPage extends Component<PageProps> {
-  static getInitialProps = async (context: GetInitialPropsProps) => {
-    const version = pageStore('toggles').apiV2 ? 2 : 1;
-    const query = context.query.query;
-    const page = context.query.page ? parseInt(context.query.page, 10) : 1;
-    const works = await getWorks({ query, page, version });
+      {works && works.results.length === 0 &&
+        <div className={`row ${spacing({s: 4}, {padding: ['top']})}`}>
+          <div className='container'>
+            <div className='grid'>
+              <div className={grid({s: 12, m: 10, l: 8, xl: 8})}>
+                <p className='h1'>
+                  We couldn{`'`}t find anything that matched
+                  {' '}
+                  <span
+                    className={classNames({
+                      [font({s: 'HNL1'})]: true
+                    })}
+                    style={{fontWeight: '400'}}
+                  >
+                    {query}
+                  </span>.
+                </p>
+              </div>
+            </div>
+          </div>
+        </div>
+      }
+    </PageLayout>
+  );
+};
 
-    if (works.type === 'Error') {
-      return { statusCode: works.httpStatus };
-    }
-
-    const pagination = PaginationFactory.fromList(
-      works.results,
-      Number(works.totalResults) || 1,
-      Number(page) || 1,
-      works.pageSize || 1,
-      {query: query || ''}
-    );
-
-    return {
-      works,
-      query,
-      page,
-      pagination: pagination,
-      version
-    };
+Works.getInitialProps = async (
+  ctx: Context
+): Promise<Props> => {
+  const query = ctx.query.query;
+  const page = ctx.query.page ? parseInt(ctx.query.page, 10) : 1;
+  const filters = ctx.query.toggles.unfilteredCatalogueResults ? {} : {
+    workType: ['q', 'k'],
+    'items.locations.locationType': ['iiif-image']
   };
 
-  handleSubmit = (event: SyntheticEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    // $FlowFixMe: Find out how to access specific `elements` on ``HTMLFormElement`
-    const queryString = event.currentTarget.elements.query.value;
+  const worksOrError = query && query !== '' ? await getWorks({ query, page, filters }) : null;
 
-    // Update the URL, which in turn will update props
-    // $FlowFixMe
-    Router.push(worksLink({ query: queryString, page: 1 }).href);
-  }
+  return {
+    initialPage: page,
+    initialWorks: worksOrError,
+    initialQuery: query,
+    filters
+  };
+};
 
-  beaconV2 = async () => {
-    if (pageStore('toggles').beaconV2) {
-      const {query, page} = this.props;
-      await getWorksV2({
-        query: query || '',
-        page: page || 1,
-        filters: {
-          workType: ['q', 'k'],
-          'items.locations.locationType': ['iiif-image']
-        }
-      });
-    }
-  }
-
-  componentDidMount = async () => {
-    this.beaconV2();
-  }
-
-  componentDidUpdate = async () => {
-    this.beaconV2();
-  }
-
-  render() {
-    return (
-      <Works
-        version={this.props.version}
-        page={this.props.page}
-        query={this.props.query}
-        works={this.props.works}
-        pagination={this.props.pagination}
-        handleSubmit={this.handleSubmit}
-      />
-    );
-  }
-}
-
-export default WorksPage;
+export default Works;
