+++ catalogue/webapp/services/catalogue/links.js
@@ -28,32 +28,6 @@ export function workLink({ id, query, page }: WorkLinkProps): LinkProps {
   return {
     href: {
       pathname: `/work`,
-      query: { id, query, page }
-    },
-    as: {
-      pathname: `/works/${id}`,
-      query: removeEmpty({ query, page })
-    }
-  };
-}
-
-export function worksLink({ query, page }: WorksLinkProps): LinkProps {
-  return {
-    href: {
-      pathname: `/works`,
-      query: { query, page }
-    },
-    as: {
-      pathname: `/works`,
-      query: removeEmpty({ query, page })
-    }
-  };
-}
-
-export function workV2Link({ id, query, page }: WorkLinkProps): LinkProps {
-  return {
-    href: {
-      pathname: `/workv2`,
       query: {
         id,
         ...removeEmpty({
@@ -63,7 +37,7 @@ export function workV2Link({ id, query, page }: WorkLinkProps): LinkProps {
       }
     },
     as: {
-      pathname: `/worksv2/${id}`,
+      pathname: `/works/${id}`,
       query: removeEmpty({
         query: query || undefined,
         page: page && page > 1 ? page : undefined
@@ -72,17 +46,17 @@ export function workV2Link({ id, query, page }: WorkLinkProps): LinkProps {
   };
 }
 
-export function worksV2Link({ query, page }: WorksLinkProps): LinkProps {
+export function worksLink({ query, page }: WorksLinkProps): LinkProps {
   return {
     href: {
-      pathname: `/worksv2`,
+      pathname: `/works`,
       query: removeEmpty({
         query: query || undefined,
         page: page && page > 1 ? page : undefined
       })
     },
     as: {
-      pathname: `/worksv2`,
+      pathname: `/works`,
       query: removeEmpty({
         query: query || undefined,
         page: page && page > 1 ? page : undefined
