+++ common/styles/works.scss
@@ -43,6 +43,7 @@ $is-next: true;
 @import 'components/footer_nav';
 @import 'components/footer_social';
 @import 'components/header';
+@import 'components/highlighted_heading';
 @import 'components/image_viewer';
 @import 'components/input';
 @import 'components/lazyload';
